feat(aggregation): add unit tests
