+++ antarest/study/web/output_blueprint.py
@@ -20,6 +20,7 @@ from fastapi import APIRouter, Depends, File, Query, Request
 from starlette.responses import FileResponse
 
 from antarest.core.config import Config
+from antarest.core.serde.matrix_export import TableExportFormat
 from antarest.core.utils.utils import sanitize_string, sanitize_uuid
 from antarest.core.utils.web import APITag
 from antarest.login.auth import Auth
@@ -30,7 +31,7 @@ from antarest.study.business.aggregator_management import (
     MCIndLinksQueryFile,
 )
 from antarest.study.model import ExportFormat, StudyDownloadDTO, StudySimResultDTO
-from antarest.study.storage.df_download import TableExportFormat, export_df_chunks
+from antarest.study.storage.df_download import export_df_chunks
 from antarest.study.storage.output_service import OutputService
 from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix import MatrixFrequency
 from antarest.study.storage.rawstudy.model.filesystem.root.output.simulation.mode.mcall.digest import DigestUI
